refactor: rename pkg/sonnet to pkg/jsonnet
